Remove factory function tileLayerGLColorScale.
This adds little value, since it’s just a transparent pass-through to the constructor. When bundling as an IIFE, the factory function is exposed as L.TileLayer.tileLayerGLColorScale, which doesn’t seem like the right place for it. L.tileLayerGLColorScale seems more appropriate, but it wouldn’t be easy to export to different namespaces like that using Rollup.